requirements.txt
numpy>=1.21.0
sounddevice>=0.4.0
Pillow>=8.0.0
simpleaudio>=1.0.4 